remove api, move convert to utils
