+++ llms/mlx_lm/__init__.py
@@ -1,5 +1,4 @@
 # Copyright © 2023-2024 Apple Inc.
 
-from .api import convert
-from .utils import generate, load
+from .utils import convert, generate, load
 from .version import __version__
